+++ src/common/utils.h
@@ -6,9 +6,9 @@
 
 #include "common/structures.h"
 
-double vectorlen(Vector const& vec);
+float vectorlen(Vector const& vec);
 
-double dotProduct(Vector const& a, Vector const& b);
+float dotProduct(Vector const& a, Vector const& b);
 
 bool pointInShadow(Point const& point, Point const& light, Sphere const& sphere);
 
@@ -16,7 +16,7 @@ bool pointInShadow(Point const& point, Point const& light, Plane const& plane);
 
 void normalize(Vector& vec);
 
-double distance(Point const& a, Point const& b);
+float distance(Point const& a, Point const& b);
 
 // Returns true if intersection exists
 // If two points of intersection exist closest is returned
